move txnkeys test to shared test pkg
