+++ storage/kv/test/txn.go
@@ -132,3 +132,55 @@ func TestTxnSimple(t *testing.T, ctx context.Context, b kv.BucketTxnBucket) {
 		t.Fatal(err)
 	}
 }
+
+func slicesEqual[T comparable](a, b []T) bool {
+	if len(a) != len(b) {
+		return false
+	}
+	for i := range a {
+		if a[i] != b[i] {
+			return false
+		}
+	}
+	return true
+}
+
+func TestKVTxnKeys(t *testing.T, ctx context.Context, b kv.KeysPrefixTraversingBucketTxnBucket) {
+	err := b.Set(ctx, "hello", []byte("dlrow"))
+	if err != nil {
+		t.Fatal(err)
+	}
+	bt, err := b.BeginKeysPrefixTraversingBucketTxn(ctx)
+	if err != nil {
+		t.Fatal(err)
+	}
+	err = bt.Set(ctx, "hello", []byte("world"))
+	if err != nil {
+		t.Fatal(err)
+	}
+	// make sure we have what we set in the txn
+	keys := kv.AllKeys(ctx, bt)
+	if want, have := []string{"hello"}, keys; !slicesEqual(want, have) {
+		t.Errorf("want: %v, have: %v", want, have)
+	}
+	// delete the key
+	err = bt.Delete(ctx, "hello")
+	if err != nil {
+		t.Fatal(err)
+	}
+	// check that we don't see it
+	keys = kv.AllKeys(ctx, bt)
+	if want, have := []string{}, keys; !slicesEqual(want, have) {
+		t.Errorf("want: %v, have: %v", want, have)
+	}
+	// roll it back
+	err = bt.Rollback(ctx)
+	if err != nil {
+		t.Fatal(err)
+	}
+	// check that we don't see it in the parent store
+	keys = kv.AllKeys(ctx, b)
+	if want, have := []string{"hello"}, keys; !slicesEqual(want, have) {
+		t.Errorf("want: %v, have: %v", want, have)
+	}
+}
